requirements.txt
python-telegram-bot==21.10
fastapi==0.115.12
uvicorn[standard]==0.34.2
openai==1.78.1
sqlalchemy[asyncio]==2.0.49
aiosqlite==0.21.0
pydantic-settings==2.9.1
python-dotenv==1.1.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
